requirements.txt
setuptools
jsonschema
grpcio
grpcio-tools
requests
python-magic